pyproject.toml
[tool.poetry]
name = "delitant"
version = "0.1.0"
description = ""
authors = ["timonzet <prokoffev1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.2"
flask-sqlalchemy = "^3.0.3"
flask-login = "^0.6.2"
flask-migrate = "^4.0.4"
openpyxl = "^3.1.2"
xlrd = "^2.0.1"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
